Original Matlab File
This file was the original file created to run the orthosis robot in the data acquisition mode. This file was created with Matlab 2013a but did not function properly once used in May 2017. This file uses data output blocks to save the output data to the workspace. The data did not save properly to the workspace with only 64 data points from each signal being saved.