Old version of file
This file was updated to use the 'To Workspace' blocks in Matlab to save the output data to the workspace. This file was updated to work with Matlab 2016b. While this file did work properly it was not a robust solution. The file would have to be rebuilt as after 7-10 runs the 'To Workspace' blocks would stop outputting data.